Move code from meta/Dashboard to meta/Parameter
